Extract oauth related code
